+++ src/main/php/lang/ast/Parse.class.php
@@ -406,13 +406,24 @@ class Parse {
         while ('}' !== $this->token->symbol->id) {
           $this->token= $this->advance();
           $class= $import.$this->token->value;
-          $types[]= $class;
           $this->scope->import($class);
           $this->token= $this->advance();
+          if ('as' === $this->token->value) {
+            $this->token= $this->advance();
+            $types[$class]= $this->token->value;
+            $this->token= $this->advance();
+          } else {
+            $types[$class]= null;
+          }
         }
         $this->token= $this->advance();
+      } else if ('as' === $this->token->value) {
+        $this->token= $this->advance();
+        $types= [$import => $this->token->value];
+        $this->token= $this->advance();
+        $this->scope->import($import);
       } else {
-        $types= [$import];
+        $types= [$import => null];
         $this->scope->import($import);
       }