+++ src/test/php/lang/ast/unittest/parse/NamespacesTest.class.php
@@ -14,14 +14,24 @@ class NamespacesTest extends ParseTest {
 
   #[@test]
   public function use_statement() {
-    $this->assertNodes([['use' => ['lang\ast\Parse']]], $this->parse('use lang\ast\Parse;'));
+    $this->assertNodes([['use' => ['lang\ast\Parse' => null]]], $this->parse('use lang\ast\Parse;'));
+  }
+
+  #[@test]
+  public function use_with_alias() {
+    $this->assertNodes([['use' => ['lang\ast\Parse' => 'P']]], $this->parse('use lang\ast\Parse as P;'));
   }
 
   #[@test]
   public function grouped_use_statement() {
     $this->assertNodes(
-      [['use' => ['lang\ast\Parse', 'lang\ast\Emitter']]],
+      [['use' => ['lang\ast\Parse' => null, 'lang\ast\Emitter' => null]]],
       $this->parse('use lang\ast\{Parse, Emitter};')
     );
+  }
+
+  #[@test]
+  public function grouped_use_with_alias() {
+    $this->assertNodes([['use' => ['lang\ast\Parse' => 'P']]], $this->parse('use lang\ast\{Parse as P};'));
   }
 }